+++ pyplanet/apps/contrib/mx/templates/search.xml
@@ -1,0 +1,15 @@
+{% extends "mx/toolbarlist.xml" %}
+
+
+  {% block toolbar %}
+  <entry pos="2 -4.25" size="60 6" z-index="2" default="{{ search_map|default('Search Map...', True) }}"
+             focusareacolor1="0003" name="map" selecttext="1" textsize="1.6"
+             focusareacolor2="FFF2" scriptevents="1" valign="center" />
+  <entry pos="63 -4.25" size="60 6" z-index="2" default="{{ search_author|default('Search Author...', True) }}"
+             focusareacolor1="0003" name="author" selecttext="1" textsize="1.6"
+             focusareacolor2="FFF2" scriptevents="1" valign="center" />
+
+  <label pos="124 -1.35" size="6 6" focusareacolor2="FFF2" focusareacolor1="0003"
+         text="🔍" action="{{ id }}__mx_search" />
+
+  {% endblock %}
